fix(Dashboard Layout): Move layout folder to dashboard folder to have a better structure
