+++ src/app/dashboard/dashboard-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { LayoutComponent } from '../layout/layout.component';
+import { LayoutComponent } from '../dashboard/layout/layout.component';
 import { PatientsComponent } from './patients/patients.component';
 import { PatientDetailComponent } from './patients/patient-detail/patient-detail.component';
 import { OrderDetailComponent } from './orders/order-detail/order-detail.component';
