+++ src/app/dashboard/layout/layout.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { LayoutComponent } from './layout.component';
-import { SharedModule } from '../shared/shared.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
     declarations: [LayoutComponent],
